LL-220 #close Functionality implemented.
The CSS part might probably be made better but the general logic should
be working well. Added isUsed to Vocabulary and will use that to signal
that entity is used within some of the organize type widget (not only
canvas in some future release). Added new default image for entity (a
link one is being used as any untyped entity is a URI).