Initial changes for new API verssion.
This is still a non-working version. There is a new logic for user
names, that is still hard-coded along with new key (will need to
substitute with a better solution). The timeline will not be loaded aftr
a couple of refreshes as the currently selected episode version and
timeline states for episode versions are not set. There also seems to be
an issue with a user entity being added twice (and the latter one is
being assigned all the episodes, this is why none are shown). Possibly
some other problems also exist.