More chabges to get things working with new API.
Current version allows user to add new episodes and versions. It also
allows a user to add new circles to esisode version. Renaming circles
also works. Changing dimensions of a circle is also functional. Changing
position of a circle (dragging) does not chnage the circle itself but
creates a new one. Creating new episide from current does not work, an
empty episode with empty versions is created. It might be that some of
the code changes introduced might be to blame.