Change to Bit toolbar and some fixes to circle and entity add.
Bit toolbar will not try to fetch an entity, that should be done from
the search or from episode load side (this creates a really strage
behavior). Added a couple of fixes for newly added esisode versons. For
some reason the object being added there is using namespace URI for keys
instead of just normal strings. Was unable to find the root of the
problem, thus instroduced the fix (the root of the problem has to be
found and fixed, all fixes have comments with XXX in front, this way it
would be easy to find them and remove when fix in sntroduced). Current
version has a strage behavior. When adding entities or circles, then
making a refresh, and then trying to clikc antity or circle (a creation
event is fired for some reason). It all works fine for newly added ones
until the refresh.