Resolved organize circle and entity problems.
Clicking on circle or entity on organize should not longer be creating
new entities. Dragability and resize seems to be working for the circle.
Entities are draggable and can be added to circles or removed from them.
Entities are also draggable when dragging the circle. Also fixes a
problem with timeline widget not having identifier set (@subject being
set to undefined).