Added a tagcloud logic to search toolbar.
The implementation could be made more general. The code dealing with
tagcloud generation could be placed into a standalone view. The logic of
waiting for entities to be loaded with tags before showing tagcloud
using _.defer is not that good. A better solution might be needed that
would change the tagcloud in case any of the models for currently found
entities changes.