Added some code to get episode toolbar displayed.
Currently using a custom event emitted by EpisodeManagerView that is
captured by AppView, then current version is being set as a model for
the EpisodeToolbarView (this is probably wrong and episode should be
sent there instead, or an episode model should be extracted from the
version and used as a model instead). There is a possibility that the
logic will chnage drasticly, so current code should be considered a
placeholder. Using that custom event would allow episode toolbar to emit
one with other some other viewContext (for example ToolvarView will tap
into event and set itself as a context; this way AppView will either
notify ToolbarView or EpisodeManager view of the change; if that is
needed).