LL-291 LL-238 Initial episode sharing.
Initial episode sharing is working to some extent. Both sharing for
co-editing and sharing as a copy are possible. Searching for users is
not there yet and one would need to put the single URI to the
corresponding field. Sharing as a copy does not yet allow to select
entities and circles to be shared (excluded). In case of sharing as a
copy the timeline is not present (there possibly is some problem with
that).