LL-238 LL-291 #close Selecting only certain bits implemented.
Server side is working perfectly. The visuals are probably not what they
will be in the long run, but using multiselect to tell the system which
bits should be available after the copy is working just fine. It might
still need some testing but it seems to be working (selecting the bits
that should be shared on the organze widget itself would be the best,
but that would take a lot more time to give it a rewrite).