LL-293 #close Episode search implemented.
Current implementation uses episodes that belong to current user. Label
and description fields are used to make the search (regexp; case
insensitive). Search results will be updated on enter. It could also be
changed to work on any key being pressed, this way the results will be
updated as soon as something is typed.