LL-286 #close Importance functionality working.
Importance is part of flags logic, so flags are always loaded with
entityDescGet. The code extracts the value for importance if available
and adds that to the object being added to entities collection. The tags
key is removed from the object. The logic is also ready to handle case
of multiple importance tags, the value of the latest one is chosen.