A fix for an episode toolbar.
Added a code logic from EpsodeManagerView that should probably fix any
situations when episode has not been loaded yet, but version has already
been passed in. The logic is straight-forward. If there is a version but
still no episode entity, then just stop rendering and listen to a change
on version belongsToEpisode attribute. As soon as the episode is loaded,
the attribute should be changed and view is rendered.