LL-471 #close Added logic to force load thumbnail.
Added force load logic to BitToolbarView. Every time there is not
thumbnail and no flag that an attempt to load one has been made is true,
the view will attempt to call entityDescGet to get the thumbnail loaded.
This solution should limit the amount service calls by only making a
call once. The imformation is only loaded once a new entity is being set
as a model for the corrresponding view.