LL-405 #close Displaying a message.
A waning message will be displayed if user has not episodes yet. This
could be dismissed. The logic being used is a system-wide messages one.
Also intorduces the messages utility with helper shortcuts for different
message types. It is allowed to use HTML inside message text. Also made
sure that all the logger of the system are being initialized.