Changed method call result parameters, enforced arrays.
Chnaged parameter names for call to LearningEpisode versions. The system
was expecting something else (probably changes between versions). Also
made sure that any arrays would be sent to the Service as an Array not
comma-separated string (new version of Service requires that). Making
sure that ode is working with the latest version of SSS (work in
progress).