LL-409 #close Unresponsiveness removed.
The issue was due to tagcloud waiting for all the entities in search
results to have tags (being loaded). It turned out that in case an
entity has tags set to an empty array, the return from Model .get()
attribute call would be undefined. This way the script went into a loop.
Removed the loop possibility. NB! The requirement is that all the
entities are fully loaded by the time results are returned to the success
callback for the search method (there is a comment about possible
solutions to some of the problems that might arise in the coode of
search Service call; see EntityData for more information).