Changed UserData initialization method.
The old method used to fetchEpisodes, fetchCurrentVersion and fetchRange
(events) for every new used it sees. The first two are using method calls
that return results for currently logged in user (this way calling them
multiple times is pointless). The later one will fetch events for the
right user, but as only one timeline for currently logged in user is
shown, fetching that data seems to be rather pointless (it could be
enabled if needed), fetching multiple events and their entities is also
quite resource consuming and fetching data writhout a good reason does
not seem to be wise.