Fix for multiple calls of fixForVIE within service call.
This might hppend due to callback of multile similar calls being applied
to a single result of a single call (this way the system does not make
duplicate calls at once). But that means that the first handler might
already apply fixForVIE to the result. That way subsequent handler would
try to reapply it to a result object that has alredy been fixed, seems
to mess up the identifier, as the correct one has already been added and
the previous hey holding the value has already been removed. Added a
check to see if the result object has already undergone fixing and will
not try to apply that again (another solution would be to check for the
presence of correct identifier attribute and not do anything then; still
running the function body would just be a waste).