Changes to user search with allUsers service.
Changed EpisodeToolbarView not to fetch all the system and cache within
itself. Using UserData from now on, that should be the general source
also usable for other places. UserData will try to fetch all the system
users on load, that information could be periodically updated in needed.
The logic of fetching user data for search could be changed to cache the
result, as it requires different data structure and also excludes the
current user and system user.