LL-296 Messages sending is working now.
Message sending is working now. Still need to show some messages in case
of success or error, but am still uncertain that those should be global
(possibly some local messages that would be removed with timeout). Also
added general MessageData for sending and getting messages, that one
would still need to be changed if the result should become VIE ojects
(that is probably a good idea). Requires more changes and tunes to
finally close the issue.