LL-294 Added initial implementation.
Added new data type with entity schema. Currently message schema has
entity schemas as parent, but that might change in future. Added custom
view for messages. It might be that what is needed a one custom view for
all the types that are allowed in the activity stream. Also changes are
needed for the toolbar to allow showing only certain types, current
rendering implementation is more like a first step. Also added styles.
Probably would need a general styling for the listing views (those share
having icon on the left and content on the right within one line).