LL-796 Extended service to pass parameters to handler.
Added an extension to Service handlers that allows one to define
passThroughKeys for the call within the model. This is later on used to
determine if some additional keys should be extracted from the result
and passed to the .resolve() as a second argument. The argument is an
object that has those keys within itself. The name might not be the best
one, but it does describe the logic well enough. This is a rather
general solution and could be reused for any other calls.