LL-296 Added message in case of error.
A message will be displayed in case an error occurs and message will not
be sent. The system will disable sending any new messages in case the
message is being sent (this should prevent multiple resubmit of the same
message). It could possibly benefit from some indicator for sending the
message.