LL-796 Loading with pagination is working.
Initial logic for searching with pagination is in place and fully
functional. Code should also be able to handle cases of current results
being filtered by the tag (only results with that tag are added).
Clicking on the tag stores that value within a view as well as adds the
class to it (this way it is possible to determine results being
filtered). This class is also retained when more resources are loaded
and the tagcoud is displayed again. Code is still rather ugly and
non-optimized (one case would be that the tagcloud sizes are regenerated
for all the entities once new page is being loaded, should be using the
logic of adding some new ones and the nredisplaying the tag cloud; more
helper methods and code review are also needed).