LL-832 OpenID connect working.
This requires additional setup on SSS configuration file side
(authType should be set to oidc; plus some additional configuration of
endpoints within i5CloudConf; the latter should also be enabled). The
code is relatively clean and fully functional, the last issue to resolve
would be to display some ajaxLoader for the user and prevent login form
resubmission (if that is needed).