LL-294 Added initial implementation for recommendations.
Recommendations are really partially implemented and need a lot more
tuning. Current view is based on EntityView but does look rather bad
when being dragged, need to tune the UI for that. The recommendation
placement is according to creationTime of the recommended resources,
that is probably not what is needed. There could also be some other
problems with current implementation (needs testing).