LL-282 #close Refactored the code.
Added CircleRenameModal view, this one will render the corresponding
template. The view is initialized and rendered as part of the AppView
rendering routine. It is stored within it and is later on passed to
WidgetView as an option, this way it is passed to OrganizeView that can
reuse it for the circle lable change.