LL-294 Added more activity types to be handled.
Most of the added types have unfinished code as reliably loading the
titles of episodes, circles and entities is not yet possible. Some of
the types have been marked and possibly problematic. This is due to a
large number of those and the ability to the system to handle such
amounts of data. It might be required to separate renaming of
LearningEpisodeCircle and changing position (the latter event coukd then
be disregarded).