LL-877 Increased size.
The size should now be the same as for Timeline. The only difference is
that the font is not bold for the Organize, it could be added but did
look distracting there. Also added the same change for search results,
it now has the same size as the Timeline counterpart.