LL-877 #resolve Increased overall size a bit.
Increased overall size of the entity label. Affect both Timeline,
Search and Organize. The organize differs a bit by not being bold (it
does attract too much attention there). The size on the Organize could
be further increased to being the same as the Circle label (if provided
to be reasonable). Also increased the height of Timeline a little bit.