LL-60 #close Added new min and max zoom limits to Timeline.
Tried different values with different Timeline sizes (based on browser
window). It seems that current values are well suited for mainstream
targeted window size that is 1024 px in width. It also works well enough
for FullHD screens with windows being in fullscreen mode (although in
that case the min zoom displays quarters of an hour).