LL-965 #close Added description to Bit Toolbar.
Description handling code is a lot like the one the one for handling the
Title/Label, it might be good idea to just have one method setAttribute
that checks for a user-initiated action and is provided a vocabulary
type. Same could go for the code handling the change within the View
code, it could probably use a few differently named proxy methods that
make a call to one single method (might consider implementing that when
refactoring the code).