LL-971 Added initial implementation for the overlay.
Introduced an OrganizeOverlayView that would always be instanciated in
case of Organize Widget. This way it would be possible to enable-disable
this as needed, also it would allow to enclose the related functionality
within it. Currently it will cover any organize (please note that
droppable is also disabled from within the widget rendering code) and
provide a button to disable itself. This way it is possible to test the
visuals and functionality without making the whole Organize unusable.
The logic will change afterwards.