LL-971 Overlay for episodes that require locking.
Added and an overlay for episodes that require locking in order to be
editable (Organize functionality). The logic of the overlay is that it
is disabled by default (is added into the DOM tree but hidden). On
initialization it is checked if an episode model requires overlay and
thus needs a ock to be requested. The situation of releasing lock nor
automatically enabling/releasing locks for episodes has not been added
yet. That would probably require adding listeners for model within the
OverlayView in order to check for the changes within the some attributes
of the model and thus enable/disable the overlay as needed.