LL-971 Added model change listeners to overlay.
Overlay will register change listeners for Episode model and will try to
reevaluate the overlay on each certain model attribute change. Please
note the comments that state a need for better solution than a timeout
in case episode model not fully loaded yet. This is due to listener on
Version for episode attribute change does not work for some reason.
Added a code to add "group" to circles of an Episode model in case it is
shared for co-editing fro the first time.