LL-905 Added datepicker with icon to timeline.
The datepicker field itself is being hidden and then the visuals for
datepicker are being shown when licking on an icon. The icon is
incorrect and the visuals are or might be incomplete. Each timeline
Widget generated it's own datepicker. This way it is visually
distinctive and easier to instanciate and handle events and listeners.