LL-294 Changes to ActivityView and entity data.
Extended Service to be able to handle provided "entity" and changed the
view correspondingly. The view is using the same generalli labelled
method to find the containedEntity and containedEntityLabel. The latter
function is able to handle the case of containedEntity being a Version,
in that case there is a check for really fetching a labele of an
episode. In case episode is still not loaded a listener is being added,
once the episode becomes available the view will re-render itself.