LL-1027 #comment Added lock release on logout.
When logging out the system will check is the current episode is still
lockd by the user and will release the lock silently. The case for
closing the browser windows is not implemented yet, this would probably
have to use 'onbeforeunload' event of window, although that one is also
fired in case of a refresh. Will need to experiment with it first.