LL-1062 #comment Added incomplete simple implementation.
Please also note that addOrUpdate calls to VIE are now provided a
options with overrideAttributes set to true. This is needed to be added
everywhere so that if the attribute changes and the object is loaded
again from Service then the attributes are overwritten and not become
array of both values (this was making the Organize crash but should also
be used for other cases). The implementation is also partial and needs
to be improved (not all the cases are covered) and the clearning of
Organize within the library has to be checked and probably improved not
to leave things hanging.