LL-907 #comment Batter solution for episode overlay.
Overlay is always enabled and shown by default. Although it does check
if the model is present ruring the time of rendereing. If the model is
not yet present, the lock button will be hidden and a loading text will
be shown. Later on when the model is fetched the text will be removed,
button shown and loading text will be removed.