LL-1080 #comment Added local episode removal.
Current solution will not remove the episode from the server (need
information on the Service call for that). It would also leave some view
and data objects behind (should not be such a big deal). Had to add a
lot of code to deal with potential problems of pointers to
episode/version still being present (a better solution would be to
remove those, but that is not feasible at the moment). Needs testing.