LL-412 #resolve Changed logic for determining entity being used.
The new logic will now detect if there is a current episode and if the
entity has been used in that episode. The system will not be adding
information on entity belonging to episode into an attribute of the
entity model. This is not the fastest method and could be inspected to
make it faster.