LL-1101 #resolve Changed to sharedSpace for adding tags.
Please note that text "User's tags" is within the Bit Toolbar is no
longer valid, as these would be all the tags assigned to the bit
(removing a tag would also remove it for any other user). This is a
general logic and does not depend on how the user got access to teh bit.