LL-1082 #comment Initial possible implementation.
This is a work in progress and might need more changes. It surely needs
to be thoroughly tested. The code will force refresh the Organize if
locked/unlocked. This is due to svgjs drag library problem with
temporarily dissabling draggability. It is safer to just set the correct
flag and redraw. The view should probably not be called Overlay any more
as it no longer is one. It could probably be beneficial to abb both
buttons (acquire/release) to this new bar element, this way the code
would be in one place.