LL-1102 #comment Solved some issues.
Added code to deal with some issues. Specifically prevent some of the
attribute changes to episode lock attributes from runnin the handler
method (in case of manually clicking the acquire/release button), this
is not needed and is handled by methods themselves. Also some more
chnages to make the code run smoother. At least a rewrite to change the
name from Overlay to Bar is needed, because it no longer is an overlay.