Added a bit of logic for the new API V2.
Rewrote the authentication to fully work with API V2. Added
configuration for API V2 service location. Retwote userAll to use the
new API V2. This should be good enough example of how to handke the new
API in future, assuming a minimal change in the code. Added a chunk of
code to deal with change when AUTHOR is an object not a URI as it used
to be, it just takes the URI our of the Object provided and replaces
Object with URI (added comments in place that mention it not being
permanent solution).