Added handler for cases with data being part of the address.
This is called injectVariable within the service definition. The path
has the variable with the same name in it prefixed with colon (like:
:ytheVariable). Then it the value is passed as part of the params being
sent. The variable is being extracted from the params and injected into
the path, replacing the placeholder. This might not be the final
implementation, but should works well enough for the time being.