Removed API V1. NB! Requires changes to config if present.
Replaced HostRESTV2 as the API V1 is no longer being used. API V2 is
required to be used instead. There are a few more services to test and
issues to resolve (one of those should be file download). But all in all
most of the core functionalities are working well. There is a need to
change the Service and remove last traces of API V2 logic from it (one
of the last remining part is logout on OIDC error).