Rewrote logic for episoded.
The issue was about previous API always providing the attribute "user"
with the currently logged in user pointer. Now it is no longer present
and most of the time "author" is suitable. This time the system was
fetching the current user from it. As it was no longer so, it had to be
changed. Also rewrote the integrity check to work with hasUsers not just
the current author.