Chnaged promises handling for ActivityStream.
Using internal Deffereds and promises that always resolve even if the
initial promise returned by the call fails. This is because the code
expects all the promises to resolve before proceeding with the data. If
any of the calls failed, it showed nothing. This should not happen in
real use, but still is a sane change.