LL-1190 #comment Changed to use dblclick for circle change.
Some portions of code are still left in there to prevent some
unnecessary events being fired. Added a bit of checks and changes so
that the system would fire less CircleChanged events, mostly cases when
no change really happended (this was due to circle unselection process
doing its bidding). Although it did create a tighter connection between
Organize.js and OrganizeView, it still is rather sane and clean enough
solution. Will need to be tested, but currently everything seemed to
work well enough and it seemed that all the uunecessary chatter for
CircleChange that was not needed got removed.