LL-416 LL-1163 A solution to allow selecting covered circles.
The solution is simple enough, and is based on first clicking and
selecting the needed circle and then dragging it. When user clicks the
circle the system inspect any other circles to see if the coordinates of
clicked point could also belong into these. If there are ano other
circles that could be targetted, then the system inspects them all and
selects the one that is the smallest. It is then selected instead and is
moved in front of any others. The solution has one flaw as user is
required to select the circle first and only then start dragging,
otherwise the circle is not moved in front of all the others and the
circle that was on top would be dragged (this has to do with positions
of elements within the SVG; the one that comes first wins.