LL-416 LL-1163 Drag constraints and fixes to scrollable Organize.
It is now possible to set restrictions to draggable area, this will not
allow one to drag the circle or bit outside of visible area. There was
also a bug with determining positions and thus the whole system when the
container was scrolled. It got fixed by requireing offsets from the
parent element and not from the Organize itself (would have to be tested
for any unexpected side effects). A listener to scroll was also added
that would update internal offsets once the scrolling has stopped. This
is not a perfect solution but it should be rather good and working well.