LL-1214 #resolve LL-1154 Solved the bit locking issue.
Changed the way covered subcircle is moved to front. New solution moves
it just in front of the one covering. There is also a new logic for
determining the covered subcircle. The previous code chose the first
matching one and disregarded all others with same size. The new version
takes the one tha is positioned behind all the others that have the
same size.