LL-1234 Small check for author being entity or simply URI.
Looks like there might be a possibility that the code could be running
with episode author object not being set yet (still being just an URI).
This is why a check was added to see if that someting is an entity or
not, all is well because code needs URI anyway.