LL-1234 Changed timers for fetchData and stateSet, using lookup structure now.
The issue is that we are dealing with a singleton that handles all of
the models for TIMELINEs, this is why those were interfearing with each
other. New approach does not let that happen. NB! Also made a change to
the timelineStateSet call, it no longer changes the ID/Subject of the
TIMELINE to correspond to the new id of the current TimelineState
entity. These just keep their internal identifiers and those do not
change. This was causing issues for the lookup.