LL-1236 Added help videos plus some more changes and tunes.
All help videos were placed as subsections to correcponding sections
(called FAQ). In general the logic is to add it to the section it
belongs on the B&P UI. If it concerns multiple places in the UI, then it
was placed under the section it has the strongest relation with. Also
added a colorbox library to show embeds in dialog. Fixed URL to Evernote
Help as previous one was pointing to a resurce in German (also forced
that to open in new tab/window). Readded line breakes for the fire to be
editable, will keep it that way for now (although that also affected the
styling part).