LL-1249 #comment Solution general logic is present.
The issue is that it allows to change some parts of the view without the
full re-render. In that case it will not compensate other elements for
the size change. It just has to be dealt with be fore closing the issue.
Apart from that it should be working well enough.