LL-1249 #resolve Made sure that partial updates are considered.
The logic is to allow to change height on every partial update of the
view (those could potentially change the height of the navbar). The same
functionalities are used as part of the general render call, a flag is
set to disable running the routine multiple times as it will be
enevitably triggered in the end of the render routine anyway.