LL-1261 Fixed episodes loading, issue with TimelineData missing.
New version has TimelineData removed from Episode Version model. The
Service Caller was execting that and a failure to find one resulted in a
silent response of failed (to the whole loaded episodes). This has to be
kept in mind, as some of the successful calls (status 200) could
internally end up flagged as failed due to some components missing (in
this case it was a subResult that had to be turned into object to be
later on consumed by VIE).