LL-1262 Initial implementation, still has quite a few flaws.
In addition to having a few flaws and quirks, this one did not change
the code logic much (meaning code still thinks that there could be
Timeline Widgets for Episode Version). It might evene be left as it is,
as those Widgets will never be there. This would still leave a big chunk
of obsolete logic being around. NB! The code uses setTimeout to simulate
loading, this is due to user being added before the whole system begins
(along comes the Timeline widget model). This prevents the handler from
seeing it, not sure how to fix that one yet. All in all, it is
functional.