LL-1263 Initial implementation, moved away from events.
This implementation uses pure entities without messing with events. The
codebase was also changed to reflect that and all (most) usages of
events and corresponding namings were eradicated. The events are still
left as they are, although any trace of those could be removed in the
long run. Current implementation is still pending to get all the needed
changes from the API (a few parameters are still missing and there is a
way to deal with pagination, loading paged data does not make any sense
in case of timeline).