LL-1263 Enabled additional flags on a service.
Enabled flags to load tags and flags, also added pageSize flag as a
default with value of zero (this will disable pagination and load all
results at once). Swapped type of "file" to "uploadedFile" as it was
causing issues. This also affects search, although search seemed to be
functioning as it used to.