LL-1263 Clean-up and code overiew, mostly.
In addition to that, a change to Timeline Widget model creation and
rendering was added. As it is just one instance and it does not really
depend on anything, all of the functionality is handled by the AppView
instead of creating in AppData and then listening to the creation in
AppView (the issue is that initial user is created before AppView is
initialized, thus it is impossible to listen to its add event). AppView
initialization deal with model initialization and rendering just expects
it to be present. Seems like a suitable solution for the case.