LL-1157 Added initial visuals for Bit Create / File Upload.
All the UI elements are correct and fuly functional. The service part is
still missing (anything related to that is still not fully correct).
Also added a check for FormData being supported, show the menu visuals
as disabled and show erro message if user clicks on it.
In addition, cleaned-up the code a bit by removing a few unused
functionalities from the EpisodeManagerView.