LL-1326 #resolve Added protection to active inputs.
The protection will make sure to check the inputs and trigger update if
needed. It will postpone update and create a small lag of 150
milliseconds before switching the model. This will give the code enough
time to trigger the update. One additional measure would be to prevent
the visual effect if update fails, there is an internal flag that is
responsible for it.