LL-1314 #resolve Added logic to automatically load new entities.
This is complex solution and could have some potential small issues
(like possibility that order of entities in cluster changes after
refresh). But that is not a very high price to pay for making sure it
always works. It could be improved in the future, if the issues with
fetchData() method using some rather complicated and sometimes strange
logic is resolved.