LL-1337 #resolve Resolved issue, added more handling.
The issue was not what it was anticipated to be. This was about internal
list (so to say) of views that were being kept there for the ordering
purposes. Views for deleted episodes were not removed. Introduced
removal process. Also added a protection against episode deletion until
any previous calls to the same procedure have returned (success or
failure). The delete button is not only visually disabled, but system
also checks that internally.