LL-1339 Changed logic for Version as attached entity.
Removed logic that assumed that Version and Episode would be linked
eventually (it might not happen in context of activities). New logic
assumes that in case Version is found within "entity" attribute, Episode
should be within an array of entities of "entities" attribute. This
simplifies the logic and should be rather fast. It also removed a notion
of waiting until linking of Episode and Version happens, will never
happen for objects that are not within active episodes of current user.