LL-1358 #resolve Added fix, also fixed bigger issue.
A fix for current issue was to check if the variable has string within
it, only then apply a replacer with regular expression. The bigger issue
and the cause of the problem turned to be the VIE turning attribute
having URI of an object within a collection into that object (which is
the general logic). A special configuration option and handler code was
introduced within the VIE.js itself. It allows to set some of the
attirbutes not to be considered useful for returning the linked Entity
(even if one exists within the collection), this would instead be
treated as a special case and value would always be rerutned as-is by
the get/previous methods.